windows compile support
